talkapp/models.py
from django.db import models
from django.contrib.auth.models import User
from django.utils import dateformat
from django.conf import settings
from django.contrib.auth.hashers import make_password
from phonenumber_field.modelfields import PhoneNumberField

class UserAdditionals(models.Model):
    """
    Модель для дополнительных настроек пользователя.
    """
    user = models.ForeignKey(to=User, on_delete=models.CASCADE)
    phone_number = PhoneNumberField()

# Create your models here.
