requirements.txt
Django
redis
channels
Pillow
redis-server
django-cleanup
channels-redis
gunicorn
django-storages
boto3
django-debug-toolbar
django-phonenumber-field[phonenumbers]
